fixes according to reviews
