+++ packages/client/src/game_modules/Player.ts
@@ -1,6 +1,6 @@
 import { BaseObject } from './MovableObject';
 import { Scene } from './Scene';
-import { Size, Position, CustomWindow } from './types';
+import { TSize, Position, CustomWindow } from './types';
 
 import JumpSprite from '../assets/jump.png';
 import RunSprite from '../assets/run.png';
@@ -21,7 +21,7 @@ export class Player extends BaseObject {
 
   private readonly image: HTMLImageElement;
 
-  constructor(position: Position, size: Size) {
+  constructor(position: Position, size: TSize) {
     super(position, size);
 
     this.verDelta = 0;
