+++ packages/client/src/types/game.interface.ts
@@ -2,5 +2,5 @@ import { EGameStatus } from '../enums/gameStatus.enum';
 
 export interface IGame {
   status: EGameStatus,
-  score: number
+  score: number,
 }
