+++ packages/client/src/utils/apiTransformers.ts
@@ -1,9 +1,3 @@
 import { LeaderDataDTO, LeaderData } from '@src/types/leaders';
 
-export const transformLeader = (data: LeaderDataDTO): LeaderData => {
-    return {
-        id: data.id,
-        user: data.login,
-        score: data.score,
-    };
-};
+export const transformLeader = ({ id, login, score }: LeaderDataDTO): LeaderData => ({ id, score, user: login });
